Post Chat with Jimmy - This should be good for check in.
Reoriented and sized the water plane for better viewing.

Fixed issue with creating an extra set of indices on the water plane

Commented out Zoom function but kept alternate constructor for the water plane. Nothing is calling it but i may want to set up my water scene with a custom camera in the future.