Light Component added.
Had issue getting it into the game so it is in the framework.

Not quite sure how multiple lights show up through the GameObject but they do....

TODO: Inspector window for adjusting lights in real time
TODO: specular lighting
TODO: message Jimmy about the lights. The PM3 document confuses me with the wording. It says to be able to move THE game object but then also says program the SCENES lights. Is a light part of a game object or part of a scene? OR are the lights its own game object in the scene. Right now the lights are attached to an existing game object.