NEW WORKING COMMIT FOR PROJECT MAINTENANCE 3
Removed the transform options from the lighting object as they did nothing and would probably just confuse people. Select the light from the list of lights and the new menu that appears will allow you to fiddle with that light. This is my attempt to make testing easier.

P.S. My first light is an overhead white light, it can be modified like the rest and i personally like having it there to brighten up the scene. If you want to test it you can either bring the light down or bring the sphere up using the inspector.